# openspec/openapi-v1.yaml
openapi: 3.1.0
info:
  title: AstrBot OpenAPI v1
  version: 0.1.0
  description: >
    Target REST contract for migrating AstrBot HTTP APIs to /api/v1.
    Dynamic AstrBot configuration payloads are intentionally modeled as open
    JSON objects because their schemas are provided at runtime by template
    endpoints.
    Developer API keys currently support these scopes only: bot, provider,
    persona, im, config, chat, kb, memory, data, file, plugin, mcp, skill. The
    config scope also grants bot and provider access.
servers:
  - url: http://localhost:6185
    description: Local AstrBot server
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/v1/files:
    post:
      tags: [Files]
      summary: Upload a file
      operationId: uploadFile
      x-astrbot-scope: file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/FileUploadRequest'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
...
  /api/v1/backups:
    get:
      tags: [Backups]
      summary: List backups
      operationId: listBackups
      x-astrbot-scope: system
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
    post:
      tags: [Backups]
      summary: Export a backup
      operationId: createBackup
      x-astrbot-scope: system
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BackupExportRequest'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/v1/backups/upload:
    post:
      tags: [Backups]
      summary: Upload a backup file
      operationId: uploadBackup
      x-astrbot-scope: system
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/BackupUploadRequest'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/v1/backups/upload/init:
    post:
      tags: [Backups]
      summary: Initialize chunked backup upload
      operationId: initBackupUpload
      x-astrbot-scope: system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BackupUploadInitRequest'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/v1/backups/upload/chunk:
    post:
      tags: [Backups]
      summary: Upload a backup chunk
      operationId: uploadBackupChunk
      x-astrbot-scope: system
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/BackupChunkUploadRequest'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/v1/backups/upload/complete:
    post:
      tags: [Backups]
      summary: Complete chunked backup upload
      operationId: completeBackupUpload
      x-astrbot-scope: system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BackupUploadSessionRequest'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/v1/backups/upload/abort:
    post:
      tags: [Backups]
      summary: Abort chunked backup upload
      operationId: abortBackupUpload
      x-astrbot-scope: system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BackupUploadSessionRequest'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/v1/backups/tasks/{task_id}:
    get:
      tags: [Backups]
      summary: Get backup task progress
      operationId: getBackupProgress
      x-astrbot-scope: system
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/v1/backups/{filename}:
    get:
      tags: [Backups]
      summary: Download a backup
      operationId: downloadBackup
      x-astrbot-scope: system
      parameters:
        - $ref: '#/components/parameters/Filename'
      responses:
        '200':
          description: Backup archive
    patch:
      tags: [Backups]
      summary: Rename a backup
      operationId: renameBackup
      x-astrbot-scope: system
      parameters:
        - $ref: '#/components/parameters/Filename'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BackupRenameRequest'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
    delete:
      tags: [Backups]
      summary: Delete a backup
      operationId: deleteBackup
      x-astrbot-scope: system
      parameters:
        - $ref: '#/components/parameters/Filename'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/v1/backups/{filename}/check:
    post:
      tags: [Backups]
      summary: Check a backup before import
      operationId: checkBackup
      x-astrbot-scope: system
      parameters:
        - $ref: '#/components/parameters/Filename'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/v1/backups/{filename}/import:
    post:
      tags: [Backups]
      summary: Import a backup
      operationId: importBackup
      x-astrbot-scope: system
      parameters:
        - $ref: '#/components/parameters/Filename'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BackupImportRequest'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    DashboardBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Dashboard session JWT
    DashboardCookieAuth:
      type: apiKey
      in: cookie
      name: astrbot_dashboard_jwt
  responses:
    Ok:
      description: Standard AstrBot success response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SuccessEnvelope'
...
  parameters:
...
    Filename:
      name: filename
      in: path
      required: true
      schema:
        type: string
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          const: ok
        message:
          type: string
        data: {}
      additionalProperties: false
...
    FileUploadRequest:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
...
    BackupExportRequest:
      type: object
      properties:
        include:
          type: array
          items:
            type: string
        exclude:
          type: array
          items:
            type: string
      additionalProperties: false
    BackupUploadRequest:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
    BackupUploadInitRequest:
      type: object
      required: [filename, total_size]
      properties:
        filename:
          type: string
        total_size:
          type: integer
          minimum: 1
      additionalProperties: false
    BackupUploadSessionRequest:
      type: object
      required: [upload_id]
      properties:
        upload_id:
          type: string
      additionalProperties: false
    BackupChunkUploadRequest:
      type: object
      required: [upload_id, chunk_index, chunk]
      properties:
        upload_id:
          type: string
        chunk_index:
          type: integer
          minimum: 0
        chunk:
          type: string
          format: binary
    BackupRenameRequest:
      type: object
      required: [new_name]
      properties:
        new_name:
          type: string
      additionalProperties: false
    BackupImportRequest:
      type: object
      properties:
        confirmed:
          type: boolean
          default: true
      additionalProperties: false
...